bring fork up to date

+++ EmmaSharp/Models/Members/RemoveMemberGroups.cs
@@ -16,7 +16,7 @@ namespace EmmaSharp.Models.Members
         /// Member ids to remove from the given groups.
         /// </summary>
         [JsonProperty("member_ids")]
-        public List<int> MemberIds { get; set; }
+        public List<long> MemberIds { get; set; }
 
         /// <summary>
         /// Group ids from which to remove the given members.
